Add sql to visualizations form and extract validation to a concern
